+++ pyzoo/zoo/examples/tensorflow/tfpark/gan/gan_model.py
@@ -1,20 +1,34 @@
-import tensorflow as tf
+#
+# Copyright 2018 Analytics Zoo Authors.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+#
 
-from zoo.pipeline.api.net.tf_optimizer import GanOptimMethod
-from zoo.tfpark.gan.gan_optimizer import GANOptimizer
+import tensorflow as tf
+from tensorflow.contrib.gan.python.losses.python.losses_impl import wasserstein_discriminator_loss, \
+    wasserstein_generator_loss
 
 ds = tf.contrib.distributions
 layers = tf.contrib.layers
 tfgan = tf.contrib.gan
-from tensorflow.contrib.gan.python.losses.python.losses_impl import wasserstein_discriminator_loss, \
-    wasserstein_generator_loss
-import tensorflow as tf
-from zoo import init_nncontext
-from zoo.tfpark import TFOptimizer, TFDataset
-from bigdl.optim.optimizer import *
-import numpy as np
 
-from bigdl.dataset import mnist
+
+def discriminator_loss_fn(real_outputs, gen_outputs):
+    return wasserstein_discriminator_loss(real_outputs, gen_outputs)
+
+
+def generator_loss_fn(gen_outputs):
+    return wasserstein_generator_loss(gen_outputs)
 
 
 def _generator_helper(
@@ -109,38 +123,4 @@ def unconditional_discriminator(img, weight_decay=2.5e-5):
     net = _discriminator_helper(img, False, None, weight_decay)
     return layers.linear(net, 1)
 
-sc = init_nncontext()
 
-
-def get_data_rdd(dataset):
-    (images_data, labels_data) = mnist.read_data_sets("/tmp/mnist", dataset)
-    image_rdd = sc.parallelize(images_data)
-    labels_rdd = sc.parallelize(labels_data)
-    rdd = image_rdd.zip(labels_rdd) \
-        .map(lambda rec_tuple: [((rec_tuple[0] / 255) - 0.5) * 2,
-                                np.array(rec_tuple[1])])
-    return rdd
-
-
-training_rdd = get_data_rdd("train")
-testing_rdd = get_data_rdd("test")
-dataset = TFDataset.from_rdd(training_rdd,
-                             names=["features", "labels"],
-                             shapes=[[28, 28, 1], []],
-                             types=[tf.float32, tf.int32],
-                             batch_size=32)
-
-opt = GANOptimizer(
-    generator_fn=lambda noise: unconditional_generator(noise),
-    discriminator_fn=lambda real_data: unconditional_discriminator(real_data),
-    generator_loss_fn=wasserstein_generator_loss,
-    discriminator_loss_fn=wasserstein_discriminator_loss,
-    generator_optim_method=Adam(1e-3, beta1=0.5),
-    discriminator_optim_method=Adam(1e-4, beta1=0.5),
-    dataset=dataset,
-    generator_steps=1,
-    discriminator_steps=1,
-    checkpoint_path="/tmp/gan_model/model"
-)
-
-opt.optimize(MaxIteration(5000))
